Factor out webview launcher into own file
