+++ install_loader/src/ui.rs
@@ -5,7 +5,7 @@ use tempfile::TempDir;
 use crate::archive_extractor::{ArchiveExtractor, ExtractionStatus};
 use crate::config::InstallerConfig;
 use crate::process_runner::ProcessRunner;
-use crate::python_launcher;
+use crate::{python_launcher, installer_webview};
 use crate::support;
 use crate::support::{AppBuilder, ApplicationGUI, NextFrameCommands};
 use crate::version;
@@ -551,49 +551,7 @@ Please download the installer to your Downloads or other known location, then ru
 		ui.text_wrapped(&self.config.sub_folder_display);
 	}
 
-	fn launch() -> wry::Result<()> {
-		use wry::{
-			application::{
-				event::{Event, StartCause, WindowEvent},
-				event_loop::{ControlFlow, EventLoop},
-				window::WindowBuilder,
-			},
-			webview::WebViewBuilder,
-		};
 
-		let event_loop = EventLoop::new();
-		let window = WindowBuilder::new()
-			.with_title("Hello World")
-			.build(&event_loop)?;
-
-		// TODO: Tell python script which port to use, OR retreive port to use from python script
-		// For now we assume python script chose port 8000, but it could choose other ports if 8000 is in use
-		let webview =
-			WebViewBuilder::new(window)?.with_url("http://127.0.0.1:8000/loading_screen.html")?;
-
-		#[cfg(debug_assertions)]
-		let webview = webview.with_devtools(true);
-
-		let webview = webview.build()?;
-
-		#[cfg(debug_assertions)]
-		webview.open_devtools();
-
-		// TODO: spawn event loop on new thread? Currently this freezes the launcher-ui
-		// Could also just close the launcher UI at this point as it's not really needed anymore.
-		event_loop.run(move |event, _, control_flow| {
-			*control_flow = ControlFlow::Wait;
-
-			match event {
-				Event::NewEvents(StartCause::Init) => println!("Wry has started!"),
-				Event::WindowEvent {
-					event: WindowEvent::CloseRequested,
-					..
-				} => *control_flow = ControlFlow::Exit,
-				_ => (),
-			}
-		});
-	}
 
 	// Start either the graphical or console install. Advances the installer progression to "InstallStarted"
 	fn start_install(&mut self, is_graphical: bool) -> Result<(), Box<dyn std::error::Error>> {
@@ -622,7 +580,7 @@ Please download the installer to your Downloads or other known location, then ru
 		// or monitor console output for startup indicator string.
 
 		// TODO: fall back to old method of launching browser if this fails
-		match InstallerGUI::launch() {
+		match installer_webview::launch() {
 			Ok(_) => {}
 			Err(e) => {
 				println!("Failed to launch webview: {e}")
